fix: ApplicationDisabled support
