+++ src/env.d.ts
@@ -4,6 +4,7 @@
 interface ImportMetaEnv {
   readonly VITE_APP_VERSION: string;
   readonly VITE_SENTRY_DSN: string;
+  readonly VITE_APPLICATION_DISABLED: string;
 }
 
 interface ImportMeta {
