+++ src/pages/index.tsx
@@ -5,7 +5,7 @@ import Head from "next/head";
 import { Container } from "@/styles/styles";
 import { WalletApiClientProvider } from "@/shared/WalletApiClientProvider";
 import WalletConnect from "@/components/screens";
-import { ApplicationDisabled } from "@/components/screens/ApplicationDisabled";
+import { ApplicationDisabled } from "@/components/ApplicationDisabled";
 import { useNavigation } from "@/hooks/common/useNavigation";
 
 export { getServerSideProps } from "@/lib/serverProps";
@@ -13,14 +13,18 @@ export { getServerSideProps } from "@/lib/serverProps";
 const Index: NextPage = ({ initialized = true }: { initialized?: boolean }) => {
   const { router } = useNavigation();
 
-  const isApplicationDisabled = Boolean(process.env.NEXT_PUBLIC_APPLICATION_DISABLED === "true");
+  const isApplicationDisabled = Boolean(
+    process.env.NEXT_PUBLIC_APPLICATION_DISABLED === "true"
+  );
 
   const { uri: rawURI, mode: rawInitialMode } = router.query;
 
   const uri = rawURI && typeof rawURI === "string" ? rawURI : undefined;
 
   const initialMode =
-    rawInitialMode === "scan" || rawInitialMode === "text" ? rawInitialMode : undefined;
+    rawInitialMode === "scan" || rawInitialMode === "text"
+      ? rawInitialMode
+      : undefined;
 
   const [isMounted, setMounted] = useState<boolean>(false);
